fix more tests
