+++ tests/test_codecell_file/test_codecell_file_warnings.xml
@@ -1,9 +1,9 @@
-<document source="mystnb_codecell_file">
+<document source="mystnb_codecell_file_warnings">
     <section classes="tex2jax_ignore mathjax_ignore" ids="a-title" names="a\ title">
         <title>
             a title
-        <CellNode cell_type="code" classes="cell">
-            <CellInputNode classes="cell_input">
+        <container cell_index="1" cell_metadata="{'load': 'mystnb_codecell_file.py'}" classes="cell" exec_count="1" nb_element="cell_code">
+            <container classes="cell_input" nb_element="cell_code_source">
                 <literal_block language="ipython3" xml:space="preserve">
                     # flake8: noqa
