+++ tests/test_nb_render.py
@@ -15,7 +15,7 @@ FIXTURE_PATH = Path(__file__).parent.joinpath("nb_fixtures")
     "line,title,input,expected", read_fixture_file(FIXTURE_PATH.joinpath("basic.txt"))
 )
 def test_render(line, title, input, expected):
-    from myst_nb.parser import nb_to_tokens, tokens_to_docutils
+    from myst_nb.parse import nb_to_tokens, tokens_to_docutils
 
     dct = yaml.safe_load(input)
     dct.setdefault("metadata", {})
@@ -35,7 +35,7 @@ def test_render(line, title, input, expected):
     read_fixture_file(FIXTURE_PATH.joinpath("reporter_warnings.txt")),
 )
 def test_reporting(line, title, input, expected):
-    from myst_nb.parser import nb_to_tokens, tokens_to_docutils
+    from myst_nb.parse import nb_to_tokens, tokens_to_docutils
 
     dct = yaml.safe_load(input)
     dct.setdefault("metadata", {})
